// my-tiktuk/src/types/feed.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"
export interface PostPartProps {
    videoUrl:               string;
}

export interface PostProp {
    id:                     string;
    secretID:               string;
    text:                   string;
    createTime:             number;
    authorMeta:             AuthorMeta;
    musicMeta:              MusicMeta;
    covers:                 Covers;
    webVideoUrl:            string;
    videoUrl:               string;
    videoUrlNoWaterMark:    string;
    videoApiUrlNoWaterMark: string;
    videoMeta:              VideoMeta;
    diggCount:              number;
    shareCount:             number;
    playCount:              number;
    commentCount:           number;
    downloaded:             boolean;
    mentions:               string[];
    hashtags:               Hashtag[];
    effectStickers:         EffectSticker[];
}

export interface AuthorMeta {
    id:        string;
    secUid:    string;
    name:      string;
    nickName:  string;
    verified:  boolean;
    signature: string;
    avatar:    string;
    following: number;
    fans:      number;
    heart:     number;
    video:     number;
    digg:      number;
}

export interface Covers {
    default: string;
    origin:  string;
    dynamic: string;
}

export interface EffectSticker {
    id:   string;
    name: string;
}

export interface Hashtag {
    id:    string;
    name:  string;
    title: string;
    cover: string;
}

export interface MusicMeta {
    musicId:       string;
    musicName:     string;
    musicAuthor:   string;
    musicOriginal: boolean;
    musicAlbum:    string;
    playUrl:       string;
    coverThumb:    string;
    coverMedium:   string;
    coverLarge:    string;
    duration:      number;
}

export interface VideoMeta {
    height:   number;
    width:    number;
    duration: number;
}